Conflicts:
	cmake/HPX_SetupAllocator.cmake
	docs/manual/build_system/cmake_variables.qbk
	src/components/binpacking_factory/CMakeLists.txt
	src/components/distributing_factory/CMakeLists.txt

+++ examples/random_mem_access/random_mem_access/random_mem_access.hpp
@@ -70,7 +70,7 @@ namespace hpx { namespace components
         /// Query the current value of the random_mem_access
         int query()
         {
-            return this->base_type::query(get_gid());
+            return static_cast<int>(this->base_type::query(get_gid()));
         }
 
         /// Asynchronously query the current value of the random_mem_access
